move utils to its own package
